updated method for conditionally making sym links

+++ scripts/programs/vsc.sh
@@ -1,12 +1,16 @@
 #!/bin/bash
 
 # https://code.visualstudio.com/docs/setup/linux
+if [ ! -f /etc/apt/trusted.gpg.d/microsoft.gpg ]; then
+  curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
+  sudo install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
+fi [ ! -f /etc/apt/sources.list.d/vscode.list]; then
+  sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
+  sudo apt install -y apt-transport-https
+  sudo apt update
+fi
+
 echo "⌨️  Installing VSCode"
-curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
-sudo install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
-sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
-sudo apt install -y apt-transport-https
-sudo apt update
 sudo apt install -y code
 rm microsoft.gpg
 function install {
